Add support for multiple guilds

+++ app/events/onGuildMemberRemove.js
@@ -1,8 +1,11 @@
 const User = require("../Models/User.js")
 
 async function run(client, member) {
-    const model = await User.findBy("id", member.user.id)
-    await model.delete()
+    const model = await User.where(`user_id = '${member.user.id}' AND guild_id = '${member.guild.id}'`)
+
+    if (model) {
+        await model.delete()
+    }
 }
 
 module.exports = run
